[VM][General] Merge upstream 2016-03-01. (Pahse 1).

+++ source/src/vm/mame/emu/cpu/i86/instrv30.c
@@ -1,6 +1,6 @@
 static void PREFIXV30(_0fpre)(i8086_state *cpustate)	/* Opcode 0x0f */
 {
-	static const uint16 bytes[] = {
+	static const unsigned bytes[] = {
 		1, 2, 4, 8, 16, 32, 64, 128, 256, 512, 1024, 2048, 4096, 8192, 16384, 32768
 	};
